+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # WIP
 * i : Code cleaning
+* i : Use [Promise](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise) return version of [chrome.notifications.create](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/notifications/create)
 * i : Write clipboard using [navigator.clipboard.writeText](https://developer.mozilla.org/en-US/docs/Web/API/Clipboard/writeText)
 
 # 2.1.0